requirements.txt
streamlit==1.35.0
deepface==0.0.83
opencv-python
numpy
Pillow
